requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
apturl==0.5.2
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.8.3
Brotli==1.1.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==42.0.5
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.1
django-allauth==0.59.0
django-anymail==9.2
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-daraja==1.3.0
django-decouple==2.1
django-email==0.1.10
django-extensions==3.2.3
django-filter==24.3
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-sendgrid-v5==1.2.2
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
duplicity==0.8.21
emoji==2.8.0
fasteners==0.14.1
fcm-django==2.0.0
filelock==3.6.0
firebase-admin==6.3.0
fonttools==4.51.0
frozenlist==1.4.1
future==0.18.2
git-filter-repo==2.34.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.2.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.7
multidict==6.0.5
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
oscrypto==1.3.0
paramiko==2.9.3
pexpect==4.8.0
pillow==10.3.0
platformdirs==2.5.1
postmark==1.0.0
proto-plus==1.23.0
protobuf==4.25.1
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydyf==0.10.0
PyGObject==3.42.1
pyHanko==0.24.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==4.2.0
pyphen==0.15.0
pypng==0.20220715.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rsa==4.9
SecretStorage==3.3.1
sendgrid==6.10.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
svglib==1.5.1
svgwrite==1.4.3
systemd-python==234
tinycss2==1.3.0
Tree==0.2.4
twilio==9.2.3
typing_extensions==4.11.0
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.13.0+ds
wadllib==1.3.6
weasyprint==61.2
webencodings==0.5.1
xdg==5
xhtml2pdf==0.2.15
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
zopfli==0.2.3
